+++ history.md
@@ -1,5 +1,5 @@
 ## 1.10.0 - TBD
-* 
+*
 
 ## 1.9.0 - 2015-12-16
 * Fixed a bug where the Response object of okhttp2 client only included the last header with a given name