+++ history.md
@@ -1,4 +1,4 @@
-## 1.10.0 - TBD
+## 1.10.0 - 2016-08-10
 * HyperfitProcessor has a new processRequest overloads that takes a super type token TypeRef argument.  This allows you to return Generics from an RESTful service entry point
 ```Page<Dog> result = processor.processRequest(new TypeRef<Page<Dog>>(){}, "url-to-entry-point-that-returns-generic-like-page-of-dog");```
 * When working with missing generic arguments for embedded resources it assumed that all embedded resource are at least HyperResource interfaces, beyond that they also will implement any interfaces as defined by the resource interface selection strategy.